added integration test file
