+++ README.md
@@ -1,6 +1,6 @@
 # Cinema APIs
 
-APIs for cinema applications including buying tickets, admin panel...
+APIs for cinema management including CRUD on cinemas, halls, movies, buying tickets, admin panel...
 
 ## I. Features
