+++ server.js
@@ -1,40 +1,4 @@
-const express = require('express');
-const dotenv = require('dotenv');
-const colors = require('colors');
-const morgan = require('morgan');
-const connectDB = require('./config/db');
-const jsonResponse = require('./middlewares/jsonResponse');
-const errorHandler = require('./middlewares/errorHandler');
-const swagger = require('./config/swagger');
-
-// Load env variables
-dotenv.config({ path: './config/.env' });
-
-// Load route files
-const cinemas = require('./routes/cinemas');
-
-// Connect to database
-connectDB();
-
-const app = express();
-
-// Body parser
-app.use(express.json());
-
-// HTTP Request logger middleware
-app.use(morgan('tiny'));
-
-// Add standard response
-app.use(jsonResponse);
-
-// Mount routes
-app.use('/api/v1/cinemas', cinemas);
-
-// Add error handler
-app.use(errorHandler);
-
-// Generate API documentation
-swagger(app);
+const app = require('./app');
 
 const port = process.env.PORT || 5000;
 
@@ -52,3 +16,5 @@ process.on('unhandledRejection', (error) => {
     // Close the server and exit process with failure (Prevent the server to crash)
     server.close(() => process.exit(1));
 });
+
+module.exports = server;
