+++ startup/logging.js
@@ -1,0 +1,5 @@
+const morgan = require('morgan');
+
+module.exports = function (app) {
+    app.use(morgan('tiny'));
+};
